requirements.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
msgpack-python==0.5.6
pyaml==18.11.0
PyYAML==3.13
requests==2.21.0
urllib3==1.24.1
Werkzeug==0.14.1
